backend/requirements.txt
# CenterBack.AI Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.29
alembic>=1.13.1

# ML
scikit-learn>=1.8.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0

# Database
supabase>=2.3.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
prometheus-client>=0.20.0

# Testing
pytest>=8.0.0
